Improve Text Update Performance
Improve performance of putting truncated text into a `<text>` element.

Computing the text width of a text element with a long textContent takes
a long time (in Chrome 49.0.2612.0; other browsers not tested).  So
rather than putting in a long string, then truncating it until it fits,
start with a short string, and grow it until almost doesn't fit.